+++ CHANGELOG.md
@@ -11,6 +11,9 @@ If you depend on the `RSpec` or `Cucumber` helpers, you will need to [manually l
 
   - [#525](https://github.com/airblade/paper_trail/issues/525) / [#512](https://github.com/airblade/paper_trail/pull/512) -
     Support for virtual accessors and redefined setter and getter methods.
+  - [#518](https://github.com/airblade/paper_trail/pull/518) - Support for querying against PostgreSQL's
+    [`JSON` and `JSONB` column types](http://www.postgresql.org/docs/9.4/static/datatype-json.html) via
+    `PaperTrail::VersionConcern#where_object` and `PaperTrail::VersionConcern#where_object_changes`
   - [#507](https://github.com/airblade/paper_trail/pull/507) - Support for opting out of saving changesets on models by choice
     when the `object_changes` column exists on the default `versions` table.
   - [#500](https://github.com/airblade/paper_trail/pull/500) - Support for passing `on: []` as an argument, with only manual
@@ -18,7 +21,7 @@ If you depend on the `RSpec` or `Cucumber` helpers, you will need to [manually l
   - [#494](https://github.com/airblade/paper_trail/issues/494) - The install generator will warn the user if the migration they are
     attempting to generate already exists.
   - [#484](https://github.com/airblade/paper_trail/pull/484) - Support for
-    [PostgreSQL's JSONB Type](http://www.postgresql.org/docs/9.4/static/datatype-json.html) for storing `object` and `object_changes`.
+    [PostgreSQL's `JSONB` Type](http://www.postgresql.org/docs/9.4/static/datatype-json.html) for storing `object` and `object_changes`.
   - [#458](https://github.com/airblade/paper_trail/pull/458) - For `create` events, metadata pointing at attributes should attempt
     to grab the current value instead of looking at the value prior to the change (which would always be `nil`)
   - [#440](https://github.com/airblade/paper_trail/pull/440) - `versions` association should clear/reload after a transaction rollback.